% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_engagement.r
\name{calculate_post_engagement}
\alias{calculate_post_engagement}
\title{Calculate the engagement score of a post, from translated data.}
\usage{
calculate_post_engagement(data_in, weight_comment = 5, weight_share = 10,
  reach_modifier = 0.8)
}
\arguments{
\item{data_in}{A data frame (or data table) consisting of 4 columns : post_id reach comments
shares}

\item{weight_comment, }{weight_share Weights for comment and share variables.}

\item{reach_modifier}{Exponent to apply to the audience size (reach)}
}
\value{
A data frame consisting of 2 columns: post_id engagement_score
}
\description{
Calculate the engagement score of a post, from translated data.
}

