% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data.r
\name{create_calculated_test_data}
\alias{create_calculated_test_data}
\title{Create 'calculated' test data set.}
\usage{
create_calculated_test_data(timeline, weight_comment = 5, weight_share = 10,
  reach_modifier = 0.8)
}
\arguments{
\item{timeline}{A data frame that describes a sequence of post-related
actions.}

\item{weight_comment, }{weight_share, reach_modifier Parameters for the
actual calculation step.}
}
\value{
A data frame consisting of two columns: post_id engagement_score.
}
\description{
Create 'calculated' test data set.
}

