% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data.r
\name{create_translated_test_data}
\alias{create_translated_test_data}
\title{Create 'translated' test data set.}
\usage{
create_translated_test_data(timeline)
}
\arguments{
\item{timeline}{A data frame that describes a sequence of post-related
actions.}
}
\description{
Create 'translated' test data set.
}

