% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data.r
\name{find_reach_comments_shares}
\alias{find_reach_comments_shares}
\title{Find reach, comments, shares.}
\usage{
find_reach_comments_shares(post, timeline)
}
\arguments{
\item{post}{A post_id}

\item{timeline}{A timeline of post-related actions, in the same form as
gloograph::test_timeline}
}
\description{
Find reach, comments, shares.
}

