% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{test_timeline}
\alias{test_timeline}
\title{Hypothetical timeline of space and timeline post actions}
\format{\describe{
  \item{time}{When the action was taken.}
  \item{user_id}{The user who took a specified action}
  \item{action}{What the user did}
  \item{object_type}{The type of object that the user took action on}
  \item{object_id}
}}
\usage{
test_timeline
}
\description{
A dataset containing a small timeline of posting, commenting, and sharing
actions taken by hypothetical users. Used for testing the data processing
functions included in the gloograph package.
}
\keyword{datasets}

