% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_data.r
\name{create_random_test_timeline}
\alias{create_random_test_timeline}
\title{Create random test timeline.}
\usage{
create_random_test_timeline(seed = NULL)
}
\arguments{
\item{seed}{Optional random seed to reproduce results.}
}
\value{
A test timeline data frame of the same form as 
gloograph::test_timeline
}
\description{
Create random test timeline.
}

