% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{test_timeline}
\alias{test_timeline}
\title{Hypothetical timeline of space and timeline post actions}
\format{\describe{
  \item{time}{When the action was taken.}
  \item{subject}{The user who took a specified action}
  \item{action}{What the user did}
  \item{object}{Who the user did the action to}
  \item{post_id}{In the case of a post oject, the id of the post}
}}
\usage{
test_timeline
}
\description{
A dataset containing a small timeline of posting, commenting, and sharing
actions taken by hypothetical users. Used for testing the data processing
functions included in the gloograph package.
}
\keyword{datasets}

